requirements.txt
# Data Downloading
kagglehub>=0.2.0

# Core ML & Computer Vision
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python>=4.8.0

# Data processing
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
Pillow>=10.0.0
scikit-image>=0.21.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
loguru>=0.7.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Development (optional)
jupyter>=1.0.0
ipython>=8.0.0
black>=23.0.0
flake8>=6.0.0